Add entities and intents files
In this folder, there are files for each Entity and each Intent the bot can perform (two separate folders). Be careful when you modify the files and do not touch the location, python reads directly from them.

If you come up with new entities and intents please respect the format of the txt. Each new entity or intents must be on a single line. 
Intents: Use " [ ] " as a placeholder for entities and tag in " ( ) " they type of entity.
Entities: Be sure to have understood where the file gets feed into before adding new entities that might be not pertinent to the file